# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SnookerPocket API with Swagger",
    "description": "api description",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/seizoenen/{id}": {
      "get": {
        "tags": [
          "Seizoen"
        ],
        "summary": "get one seizoenen by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Seizoen"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Seizoen"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "bevriesKlassement": {
                  "example": "any"
                },
                "klassement": {
                  "example": "any"
                },
                "speeldagen": {
                  "example": "any"
                },
                "startdatum": {
                  "example": "any"
                },
                "seizoenBeeindigd": {
                  "example": "any"
                },
                "aantalJokers": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/seizoenen/": {
      "get": {
        "tags": [
          "Seizoen"
        ],
        "summary": "get all seizoenen",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Seizoen"
        ],
        "summary": "add a new seizoen",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Add new seizoen.",
            "schema": {
              "$ref": "#/definitions/AddSeizoen"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/seizoenen/{id}/speeldagen": {
      "get": {
        "tags": [
          "Seizoen"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Seizoen"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "schiftingsantwoord": {
                  "example": "any"
                },
                "schiftingsvraag": {
                  "example": "any"
                },
                "wedstrijden": {
                  "example": "any"
                },
                "speeldagVotes": {
                  "example": "any"
                },
                "klassement": {
                  "example": "any"
                },
                "eindDatum": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/seizoenen/{id}/klassement": {
      "get": {
        "tags": [
          "Klassement"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/seizoenen/{id}/latestklassement": {
      "get": {
        "tags": [
          "Klassement"
        ],
        "summary": "get latest klassement",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/speeldagen/": {
      "get": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "schiftingsvraag": {
                  "example": "any"
                },
                "schiftingsantwoord": {
                  "example": "any"
                },
                "wedstrijden": {
                  "example": "any"
                },
                "speeldagVotes": {
                  "example": "any"
                },
                "klassement": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/api/speeldagen/{id}": {
      "get": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/speeldagen/{id}/klassement": {
      "get": {
        "tags": [
          "Klassement"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          }
        }
      },
      "post": {
        "tags": [
          "Klassement"
        ],
        "summary": "update klassement",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/wedstrijden/": {
      "get": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "datum": {
                  "example": "any"
                },
                "resultaat": {
                  "example": "any"
                },
                "thuis": {
                  "example": "any"
                },
                "uit": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/api/wedstrijden/{id}": {
      "get": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "datum": {
                  "example": "any"
                },
                "resultaat": {
                  "example": "any"
                },
                "thuis": {
                  "example": "any"
                },
                "uit": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddSeizoen": {
      "type": "object",
      "properties": {
        "bevriesKlassement": {
          "type": "boolean",
          "example": false
        },
        "startdatum": {
          "type": "string",
          "example": "2024-03-22"
        },
        "seizoenBeeindigd": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
  }
}